[UBI] - Integration of cryptographic features
Currently, a ciphering process is available with no further checking
on a volume basis. The ioctl() command _UBI_IOCSETVOLKEY is available
to set the key for a given volume device.
The implementation of a more advanced cryptographic support is being built.
This will eradicate some security flaws and allow the system to automatically
check the volume integrity. This will add an extra header to the UBI logical
eraseblock. Depending on the underlying UBI volume (is it virgin or not ?)
the system will find out if it can or has to turn on
this feature.

[UBI] - Integration of UBI cryptographic tools [backport to 2.6.32]

This commit introduces the basis to add
on-the-fly encryption support to the UBI layer.
The new crypto subsystem consists at this time of
4 bricks.

-> cryptokmgr : The cryptographic key manager
                This brick handles the storage of cryptographic
                keys defined by the user. It will also take care
                of other functionnalities such as keeping up to date
                the validity domain of a specific key.

-> cryptounit : Crypto API object manager
                This element manages a pool of crypto API objects.
                When raw cryptographic operations are needed, such as
	        encryption/decryption or the generation of a hash,
                a unit must be retrieved from the pool.
                This layer was implemented to try to solve concurrency
                and SMP "issues".

-> cryptokval : An interval tree implementation
                This layer defines the interface to toy with an interval tree.
                The interval tree **only** stores disjoint intervals.
                This complexifies the insertion process but make the storage
                more compact. This was needed since the integer range aimed
                to be stored is [0 .. MAX_INT] and the tree could be fed with
                singletons to insert (and as a matter of fact it will happen
                many times).

-> crypto     : Cryptographic primitives for use in other UBI sub-systems
                This brick makes use of the three previous ones to supply
                cryptographic high-level related methods to other UBI
                sub-systems.
                No UBI sub-system except this one should make use of
                cryptounit, cryptokmgr and cryptokval.

The commit includes some minor changes in already existing files.
Kconfig now has an option to enable/disable the crypto sub-system.
build.c initializes and terminates the crypto sub-system when needed.
cdev.c introduces a new ioctl to set up cryptographic keys.
This ioctl is defined in ubi-user.h

[UBI] - Integration of crypto methods in eba sub-system [2.6.32 backport]

The heart of the encryption process is in the eba sub-system.
When performing only raw encryption, this is the only sub-system
which is trully making use of the crypto primitives.

[UBI] - Insertion of crypto files in the Makefile [2.6.32 backport]

The makefile now includes the required files
mapping to build the crypto sub-system.

[UBI] - Interval tree implementation for UBI

This files contains an interval RB-tree
implementation for further use in the
key update worker. These trees will store
the validity domain of a specific UBI volume

[UBI] - Fixed cryptokmgr header

Fixed structure documentation and stuff

[UBI] - Work and fixes on interval trees (cryptokval)

[UBI] - Little fixes in crypto, cryptokmgr and cdev

Removed some debugging printks and corrected programming
mistakes.

[UBI] - Tweaked encryption and interface update

The crypto engine now performs tweaked encryption
based on the physical eraseblock where the data is
being read or written to.

If the volume key is K and the underlying PEB is p,
then the bitstring K_p defined as :

K_p = K XOR p

On the other hand, the crypto sub-system primitives
have been updated for a simpler use when multiple
configuration options will be triggered (such as HMAC)

[UBI] - Interval tree implementation

Fixes on interval trees : remove and get rightmost
functions now available.

[UBI] - Interval tree

Added a forgotten case in the removal function

[UBI] - Backport fixes (3.8 -> 2.6.32)

Fixed some typos and errors