[UBI] - HMAC support integration
The HMAC support is almost ready !

The entry points are set, all the major properties of UBI
devices are updated to handle the new HMAC header.
Backward compatibility wont be broken.
The kapi function do_get_device_info still returns the usual
"leb_size", but it is used elsewhere only for
memory allocation purposes.

TODO :
-> Update fastmap.c to integrate the HMAC support
-> Implement ubi_io_read_hmac_hdr() and validate_hmac_hdr()
-> Implement the computation of the HMAC tags (in progress)
-> Implement the lookup function "by HMAC tags" for key search
-> Update the update worker

The HMAC tag computation method is still in study.
The implementation should be available as soon as the
computation pattern is clear. (Study of hash functions ...)