[UBI] - Tweak for volume key + Clean Up
From now on, the key is tweaked on a per-PEB basis.
If an operation requires encryption or decryption
at the PEB p, for a given volume encrypted with
the key K, then :

K_p = K XOR p

The XOR is made through the "crypto_xor()" function.
(See algaphi.h and the UBI crypto specification for
deeper information)

In addition, a clean up wath made to make the code clearer.
The ubi_crypto_cipher() function now takes a custom structure
as input, that gathers all the required parameters.
Doing so will prevent from having compilation flag dependencies
in the function prototype (and therefore in the calls).

The tweaked version has been tested successfully.

Incoming :
-> New branch to clean up the experimental compilation flag
   (CONFIG_UBI_CRYPTO_HMAC)
-> Further development for HMAC support.