[UBI][HMAC] - HMAC integration for UBI crypto sub-system
The Kconfig file now includes a new option to enable the
HMAC support within UBI crypto layer. This option will bring
a new LEB header and thus alter the structure of a LEB.

This has side effects that will be handled in later commits.
The system will detect if it has to turn on (or not) this
feature on a UBI volume basis.

[UBI][HMAC] - Fixes on the interval trees

Corrected some silly mistakes on the interval tree
implementation.

[UBI][HMAC] - HMAC support integration

This commits introduces some major changes
brought by the new HMAC feature. This is still
experimental, and thus, enabling this option
will lead to _many_ compilation errors...

[UBI][HMAC] - HMAC related fix in cryptounit

Corrected the crypto API call to allocate an HMAC instance

[UBI][HMAC] - HMAC feature integration

This is a development commit, where several HMAC related
behaviours are implemented, braced with the appropriate
compilation flag.