[UBI][HMAC] - HMAC integration for UBI crypto sub-system
The Kconfig file now includes a new option to enable the
HMAC support within UBI crypto layer. This option will bring
a new LEB header and thus alter the structure of a LEB.

This has side effects that will be handled in later commits.
The system will detect if it has to turn on (or not) this
feature on a UBI volume basis.