[UBI] Integration of cryptographic features
Currently, a ciphering process is available with no further checking
on a volume basis. The ioctl() command _UBI_IOCSETVOLKEY is available
to set the key for a given volume device.
The implementation of a more advanced cryptographic support is being built.
This will eradicate some security flaws and allow the system to automatically
check the volume integrity. This will add an extra header to the UBI logical
eraseblock. Depending on the underlying UBI volume (is it virgin or not ?)
the system will find out if it can or has to turn on
this feature.