[UBI][HMAC] - Integration of HMAC features
Standardization of the presence of HMAC header on
a UBI device basis. If the device is tagged as hmaced,
then all its volumes (including internal volumes such
as the layout volume) will have a HMAC header.

The HMAC header is zeroed when there is no ciphering
applied to the related LEB.