[UBI][HMAC] - Integration of HMAC features
Some problems were fixed in the key management functions.
The key finder was assuming the input HMAC header was never NULL,
which can happen when, for instance, this is the first time a LEB
is written to the Flash.

Also, the relevant code to add HMAC header handling in the
eba sub-system is almost complete. Checkups are still required
though.

Here is a little summary of what needs to be done :

-> attach :
The attaching process must be customized to initialize
the key interval trees for the different volumes found
on the device.

-> kupd :
The work begun for the key update process will be
required to make any volume consistent (i.e.
one volume = one key).

-> eba :
Finish the integration of HMAC header check-up

Of course, vtbl and vmt need a check up as well.